backend/requirements.txt
Django==2.1.1
django-auditlog==0.4.5
django-cors-headers==3.0.2
djangorestframework==3.10.1
mysqlclient==1.4.2.post1
pytz==2019.1
